TEST/uct_hello_world: added -pedantic flag to check uct.h in examples

+++ src/ucs/time/time_def.h
@@ -15,7 +15,7 @@
  * These are not necessarily aligned with metric time units.
  * MUST compare short time values with UCS_SHORT_TIME_CMP to handle wrap-around.
  */
-typedef unsigned long long   ucs_time_t;
+typedef unsigned long   ucs_time_t;
 
 
 #endif /* UCS_TIME_DEF_H */
